Added PopenBase
Added base class to facilitate extending subprocess.Popen.  PopenBase
implements the convenience methods from subprocess as classmethods so
that they are automatically available to subclasses.

For now only the convenience methods from Python 2.7 are implemented,
and the implementations make use of an added fail_on_error property.
This may be split out into a separate mixin in the future, and call()
etc implemented fully as they are in subprocess.