Removed CheckMixin
Removed mixin to Popen that checks the return code on poll() and wait().

The implementation was a little awkward - it's probably a bit odd to raise
exceptions from poll() and wait(), and it means that exceptions raised in
communicate() don't capture stdout and stderr.

All return-code checking is now handled by CompletedProcess.  The setup for
this is duplicated in run() and popen() but that could be partly abstracted
out into a helper function.