Backed out changeset 4b9b3131a376
Don't use @wraps when overriding methods as it isn't really appropriate.

The common use for @wraps is to create decorators that don't destroy the
name and docstring of functions that they decorate.

In our case, we are wrapping a specific method, so we can simply set the
correct name and docstring directly.